+++ packages/translate/src/constants.ts
@@ -1,0 +1,3 @@
+import factoryDebug from 'debug';
+
+export const debug = factoryDebug('boost:translate');